pyproject.toml
[project]
name = "binary-classification-gcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-cloud-storage>=3.1.0",
    "matplotlib==3.10.0",
    "pillow==11.1.0",
    "streamlit==1.44.1",
    "torch==2.6.0",
    "torchvision==0.21.0",
]
